backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
pypdf2==3.0.1
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
redis==5.0.1
celery==5.3.4
pytesseract==0.3.10
python-docx==1.1.0
nltk==3.8.1 